requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
appdirs=1.4.4=pyh9f0ad1d_0
audioread=2.1.9=py38h578d9bd_0
blas=1.0=mkl
brotlipy=0.7.0=py38h497a2fe_1001
bzip2=1.0.8=h7f98852_4
ca-certificates=2020.10.14=0
certifi=2020.6.20=py38_0
cffi=1.14.6=py38ha65f79e_0
chardet=4.0.0=py38h578d9bd_1
charset-normalizer=2.0.0=pyhd8ed1ab_0
cryptography=3.4.7=py38ha5dfef3_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=py_2
decorator=5.0.9=pyhd8ed1ab_0
ffmpeg=4.3.1=hca11adc_2
freetype=2.10.4=h0708190_1
gettext=0.19.8.1=h0b5b191_1005
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
idna=3.1=pyhd3deb0d_0
intel-openmp=2021.3.0=h06a4308_3350
joblib=1.0.1=pyhd8ed1ab_0
jpeg=9b=h024ee3a_2
kiwisolver=1.3.1=py38h1fd1430_1
lame=3.100=h7f98852_1001
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libflac=1.3.3=h9c3ff4c_1
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=h14aa051_19
libgfortran4=7.5.0=h14aa051_19
libgomp=9.3.0=h5101ec6_17
libidn2=2.3.2=h7f8727e_0
libllvm10=10.0.1=he513fc3_3
libogg=1.3.4=h7f98852_1
libopus=1.3.1=h7f98852_1
libpng=1.6.37=h21135ba_2
librosa=0.8.1=pyhd8ed1ab_0
libsndfile=1.0.31=h9c3ff4c_1
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libvorbis=1.3.7=h9c3ff4c_0
libwebp-base=1.2.0=h27cfd23_0
llvmlite=0.36.0=py38h4630a5e_0
lz4-c=1.9.3=h2531618_0
matplotlib-base=3.4.2=py38hcc49a3a_0
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.0=py38h42c9631_2
mkl_random=1.2.2=py38h51133e4_0
ncurses=6.2=he6710b0_1
nettle=3.7.3=hbbd107a_1
ninja=1.10.2=hff7bd54_1
numba=0.53.1=py38h8b71fd7_1
numpy=1.20.3=py38hf144106_0
numpy-base=1.20.3=py38h74d4b33_0
olefile=0.46=py_0
openh264=2.1.1=h780b84a_0
openjpeg=2.3.0=h05c96fa_1
openssl=1.1.1k=h7f98852_0
packaging=21.0=pyhd8ed1ab_0
pandas=1.2.5=py38h295c915_0
pillow=8.3.1=py38h2c7a002_0
pip=21.1.3=py38h06a4308_0
pooch=1.4.0=pyhd8ed1ab_0
pycparser=2.20=pyh9f0ad1d_2
pydub=0.25.1=pyhd8ed1ab_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pysocks=1.7.1=py38h578d9bd_3
pysoundfile=0.10.3.post1=pyhd3deb0d_0
python=3.8.5=h7579374_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytorch=1.9.0=py3.8_cuda10.2_cudnn7.6.5_0
pytz=2021.1=pyhd3eb1b0_0
readline=8.1=h27cfd23_0
requests=2.26.0=pyhd8ed1ab_0
resampy=0.2.2=py_0
scikit-learn=0.23.2=py38h0573a6f_0
scipy=1.6.2=py38had2a1c9_1
setuptools=52.0.0=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=hc218d9a_0
threadpoolctl=2.2.0=pyh8a188c0_0
tk=8.6.10=hbc83047_0
torchaudio=0.9.0=py38
torchvision=0.10.0=py38_cu102
tornado=6.1=py38h497a2fe_1
typing_extensions=3.10.0.0=pyh06a4308_0
urllib3=1.26.6=pyhd8ed1ab_0
wheel=0.36.2=pyhd3eb1b0_0
x264=1!161.3030=h7f98852_1
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
